[#19] Remove redundant code
Improve Skeleton Loading name
Update some dimension constants
Update response optional properties
