+++ lib/ui/home/home_screen.dart
@@ -7,11 +7,10 @@ import 'package:survey_flutter_ic/ui/home/home_footer_widget.dart';
 import 'package:survey_flutter_ic/ui/home/home_header_widget.dart';
 import 'package:survey_flutter_ic/ui/home/home_state.dart';
 import 'package:survey_flutter_ic/ui/home/home_view_model.dart';
-import 'package:survey_flutter_ic/ui/home/loading/home_loading_text.dart';
+import 'package:survey_flutter_ic/ui/home/loading/home_skeleton_loading.dart';
 import 'package:survey_flutter_ic/utils/dimension.dart';
 import 'package:survey_flutter_ic/ui/home/home_side_menu.dart';
 import 'package:survey_flutter_ic/ui/home/home_side_menu_ui_model.dart';
-import 'package:survey_flutter_ic/utils/loading_text.dart';
 import '../../di/di.dart';
 import '../../gen/assets.gen.dart';
 import '../../usecases/get_surveys_use_case.dart';
@@ -155,7 +154,7 @@ class HomeScreenState extends ConsumerState<HomeScreen> {
     final surveys = ref.watch(surveysStream).value ?? [];
     return surveys.isNotEmpty
         ? _homeWidget(context)
-        : const SafeArea(child: LoadingText());
+        : const SafeArea(child: HomeSkeletonLoading());
   }
 
   void _setupStateListener() {
